Shadow Apache HTTP library in assembled JAR
I think the copy of the Apache HTTP library the latis-hylatis code is
using conflicts with the version used by Spark when we run it on the
cluster (when the Spark dependency is set as "provided").

By shadowing the copy in latis-hylatis we can each use our respective
copies without conflict.

See HYLATIS-84.