Use existing SparkContext if available
The Spark REPL provides a SparkContext and trying to create a new one
will raise an exception. This uses the existing Spark context instead.

One potential issue is that some (perhaps all) of the configuration
we're applying here won't be applied to the existing SparkConf. I
don't know what is or is not applied, but so far it hasn't been an
issue.

See HYLATIS-191.